requirements/common.txt
behave==1.2.6
Flask==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
flask-restplus==0.12.1
GeoAlchemy2==0.6.2
marshmallow-sqlalchemy==0.16.2
mypy==0.700
psycopg2-binary==2.8.2
python-dotenv==0.10.1
pytz==2018.9
simplejson==3.16.0
typing-extensions==3.7.2
tavern==0.26.3
pytest==4.4.1
python-dateutil==2.8.0
